+++ client/script/game.js
@@ -6,7 +6,7 @@ import {Viewport} from 'pixi-viewport'
 import {keyboard} from '../script/PIXIutils'
 import makeFisherySprite from '../script/makeFisherySprite'
 import makeFishSprite from '../script/makeFishSprite'
-import makeMapSprite from '../script/makeMapSprite'
+import makeMapSprite from './sprites/mapSprite'
 //import {spawnFish} from '../../utilityMethods.js'
 import socket from '../socket'
 import {TILE_SIZE, SCALE} from '../script/drawMap'
