+++ client/script/sprites/index.js
@@ -1,1 +1,2 @@
 export * from './decorations'
+export * from './mapSprite'
